[M2] Update critical error
I made a critical mistake when writing updates, and I didn't check it for a long time until I needed the program myself.

The essence of the error is that when extracting strings, she tried to find CSV files (which has not yet been created) and crashed with an error, eventually creating nothing.